cryptohome: Make auth_factor.h not depend on auth_block_utility
Currently the PrepareForRemoval is implemented by calling a method on
AuthFactor, which passes the AuthBlockUtility as a parameter and calls
AuthBlockUtility::PrepareAuthFactorForRemoval on the auth block's auth
block state. This is kind of anti-pattern, and makes auth_factor.h
depend on auth_block_utility. This is inconvenient because then we can't
refer to the AuthFactor type in AuthBlockUtility/AuthBlock headers.
Remove the dependency.

BUG=b:269578937
TEST=cryptohome unittests

Change-Id: Ibbf6c65662241a9a9cac30f7b02b224a64faded0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4343794
Tested-by: Howard Yang <hcyang@google.com>
Commit-Queue: Howard Yang <hcyang@google.com>
Reviewed-by: John Admanski <jadmanski@chromium.org>