arc: arcvm-data-migrator: Set casefold option to /data/media
Files under Android's /data/media directory need to support
case-insensitive lookup, and in virtio-fs /data this has been
implemented in the virtio-fs layer. Since we're migrating some devices
from virtio-fs /data to virtio-blk /data, we need to enable the ext4
casefold feature for /data/media after the migration.

Because the casefold bit can only be set on empty directories, we need
to set it during the setup phase of a new migration. Once the bit is
set, it will be inherited by the children and will be preserved during
the migration, so all the migrated directories under /data/media will
have the same bit set (as well as the new directories created after the
migration).

This CL adds a logic to arcvm-data-migrator that creates /data/media in
the migration destination and sets the casefold bit, only when
/data/media doesn't already exist. Also adds a metrics enum value for
the setup failure.

Chrome-side CL (metrics change): CL:4349752

BUG=b:251764421
TEST=Manually create a directory /data/media/0/Pictures/foo in virtio-fs
TEST=/data and do the migration. After the migration, enter adb shell
TEST=as root and check the following:
TEST=1. `ls /data/media/0/PICTURES/FOO` works.
TEST=2. `lsattr -d /data/media/0/Pictures` and
TEST=   `lsattr -d /data/media/0/Pictures/foo` both show the 'F' bit.
TEST=3. The attributes (owners, permission, xattr, ext4 attr, quota id)
TEST=   of /data/media are set appropriately.
TEST=4. New directories created under /data/media after the migration
TEST=   also have the 'F' bit set.

Change-Id: I08f3dcd9e18931c2b883dc0111787f4bbe73bee3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4346412
Reviewed-by: Youkichi Hosoi <youkichihosoi@chromium.org>
Reviewed-by: Satoshi Niwa <niwa@chromium.org>
Tested-by: Momoko Hattori <momohatt@chromium.org>
Commit-Queue: Momoko Hattori <momohatt@chromium.org>