diagnostics: Add support status check result type
Add a new type for support status check. This type is used in those
IsXXXSupported methods. It provide more details and simplify the content
of Exception type. Exceptions are still raised if a unsupported feature
is requested. However, clients are not expected to used that for
support status check.

BUG=b:271514913
TEST=CQ

Change-Id: Ic3051232b6c7794bbc162be77e8e4a408da7050f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4344835
Reviewed-by: Bastian Kersting <bkersting@google.com>
Reviewed-by: Kerker Yang <kerker@chromium.org>
Reviewed-by: Takashi Toyoshima <toyoshim@chromium.org>
Reviewed-by: Wei-Luan Wang <weiluanwang@google.com>
Tested-by: Chung-sheng Wu <chungsheng@google.com>
Commit-Queue: Chung-sheng Wu <chungsheng@google.com>