usb_bouncer: use subsystem to filter reported errors
Currently the usb bouncer will use the DRIVER environment variable from
error uevents to determine if a report error should be attributed to the
hub or xHCI driver. While this typically works, there are cases where
the DRIVER environment variable is not set. This CL updates the USB
driver to instead use the SUSBSYSTEM environment variable to parse error
uvents. The SUBSYSTEM environment variable is a default variable in
kobject_uevent_env and will always be set by the function which sends
the uevent.

BUG=b:273377432
TEST=FEATURES=test emerge-brya chromeos-base/usb_bouncer. Manually
checed that '%E{SUBSYSTEM}' was set to "usb" or "pci" in cases
where '%E{DRIVER}' was empty.

Change-Id: Iab07d781a949be677c0d2911f0576e521a967746
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4348054
Reviewed-by: Benson Leung <bleung@google.com>
Commit-Queue: Jameson Thies <jthies@google.com>
Tested-by: Jameson Thies <jthies@google.com>
Reviewed-by: Allen Webb <allenwebb@google.com>