oobe_config: Add new methods to FileHandler
Add methods to FileHandler required for the rollback metrics
library.

Include new methods to handle the metrics file and
corresponding tests to be used in the rollback metrics library.

Include new method in FileHandlerForTesting to be able to remove
the metrics enabled file in tests.

BUG=b:261850979
TEST=FEATURES=test emerge-amd64-generic oobe_config

Change-Id: Ieea1ccac7b35a986eac72948e03af36794579a3c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4350019
Reviewed-by: Miriam Polzer <mpolzer@google.com>
Commit-Queue: Cristina Guerrero <crisguerrero@chromium.org>
Tested-by: Cristina Guerrero <crisguerrero@chromium.org>