chromeos-config: Stop printing json content to stdout
The MainTests.testIdentityTableOut test of cros_config_schema_unittest
will print out the content of the transformed json file to stdout, which
contains hundreds of lines. This is unnecessary, as the test itself
doesn't even check the content of the json file.

Stop the code from printing it to stdout, by redirecting stdout to an
io.StringIO object.

BUG=none
TEST=./run_tests.sh

Change-Id: I6e263ce3fb2857607469b71ca7529735694df1d3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4351201
Reviewed-by: Jack Rosenthal <jrosenth@chromium.org>
Tested-by: Yu-Ping Wu <yupingso@chromium.org>
Commit-Queue: Yu-Ping Wu <yupingso@chromium.org>