shill: Always clear FD watchers before ares processing
Watching a closed file descriptors is discouraged as it can lead to a
dangerous race-condition i.e. if a new fd is created with the same fd
number. For the same reason, IOHandlers can cause a CHECK crash when
dealing with a closed file descriptor.

Ares might close its fds on ares_process_fd. This patch makes it so
that the IOHandlers are always cleared prior to calling any ares
processing.

Both read and write handlers need to be reset as ares_process_fd might
close the fd due to timeouts. As this invalidates the usage of map, the
handlers are now stored as vectors.

BUG=b:218318341
TEST=tast run <DUT> network.HealthCaptivePortalHTTP.* \
     PASSED (except for portal-suspected, which is currently failing);
TEST=tast run <DUT> network.ShillCaptivePortalDNS;
TEST=FEATURES=test emerge-<BOARD> shill;
TEST=Observe that DNS queries are successfully done by shill through
     logs;

Change-Id: I5db4f780b9876a521c0c441ff24c57d2438c95cd
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4348269
Commit-Queue: Jason Jeremy Iman <jasongustaman@chromium.org>
Tested-by: Jason Jeremy Iman <jasongustaman@chromium.org>
Reviewed-by: Steven Bennetts <stevenjb@chromium.org>