cryptohome: Remove the extra EndAuthenticateSession
Biod auth session is actually designed for the caller to start the auth
session again without ending it. This is so that it can enforce the user
to lift their finger before the next attempt. Remove the extra
EndAuthenticateSession otherwise all the attempts will be instantly
consumed by a single touch that lasts a few seconds.

BUG=b:249749541
TEST=test with cryptohome CLI that auth scan signal is only sent after
the finger is lifted and put down again in a same prepare session.

Change-Id: If66fca0b8b8f359271242d9f45493e4f1e3204f7
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4349467
Reviewed-by: John Admanski <jadmanski@chromium.org>
Tested-by: Howard Yang <hcyang@google.com>
Commit-Queue: Howard Yang <hcyang@google.com>