imageloader: drop u8.
This is preparation of C++20 support in ChromeOS repository.

The test code uses u8"..." string literal, which is typed char8_t[]
in C++20, and so cannot directly convertible into std::string.

The content of th u8 string is just in ASCII range, which we assume
on using char (while it is out of the spec, though, in very precise).

Thus, just using char array is as same as what we have.

BUG=b:217226507
TEST=tryjob

Change-Id: I44476b35666f406932c2b19ba6c15c0385cdcefa
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4352189
Commit-Queue: Hidehiko Abe <hidehiko@chromium.org>
Tested-by: Hidehiko Abe <hidehiko@chromium.org>
Reviewed-by: Jae Hoon Kim <kimjae@chromium.org>