dlp: Add watchdog waiting for fanotify reply.
When fanotify access request comes, a watchdog
is started on a separate thread that ensures that
it's replied in less than a second, otherwise it
might be that DLP daemon hang and should be killed.

This is needed to prevent a possible kernel livelock
when it's waiting to DLP daemon to give access to the
files.

BUG=b:272441332
TEST=Modified daemon to not reply, checked that it's being exited.

Change-Id: Ic4e5607f014d51ceedaa285ade87a18e41cb6ad9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4354634
Reviewed-by: Aya Elsayed <ayaelattar@chromium.org>
Commit-Queue: Sergey Poromov <poromov@chromium.org>
Tested-by: Sergey Poromov <poromov@chromium.org>