libec: Support using FP_TEMPLATE command to reload
Allow zero template size so that we can simply do commit without upload
any portion of the template in the FP_TEMPLATE command.

Usually FP_TEMPLATE command commits the template along with the last
upload, but the newly added FP_PRELOAD_TEMPLATE command commits the
template by copying the preloaded template to the template buffer. By
this time the template is ready to be committed, but there are no
remaining portions that need to be uploaded anymore. So we need to
support committing with zero size. See https://crrev.com/c/3966088 that
uses this.

BUG=b:251695952
TEST=libec unittests

Change-Id: I7a787643414b8ed622a5720d643b5750bc817a2d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4357955
Reviewed-by: Craig Hesling <hesling@chromium.org>
Tested-by: Howard Yang <hcyang@google.com>
Commit-Queue: Howard Yang <hcyang@google.com>