hiberman: reduce 'hibermeta' volume size
The DM snapshots have been moved off the 'hibermeta' volume,
so the size of the volume can be reduced. The LV is now only
used for logs and metrics which require very little storage.

We still observe file system corruptions related with hibernate
whose cause aren't fully understood at this time. A larger
'hibermeta' volume seems to delay the occurrence of these issues,
so use a volume size of 1GB for now to facilitate development
without the need for constantly re-installing the OS.

BUG=b:267968644
TEST=hibernate works on brya

Change-Id: I2d1206cba38509e4bdd1f1fa7d973f918c235e5c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4401963
Tested-by: Matthias Kaehlcke <mka@chromium.org>
Commit-Queue: Matthias Kaehlcke <mka@chromium.org>
Reviewed-by: Brian Geffon <bgeffon@chromium.org>
Reviewed-by: Allen Webb <allenwebb@google.com>