missive: Move key request logic from queue to storage
Move logic to request/update the encryption key to storage. Previously the
queue checked whether we need the key and then initiate the request.
This change is necessary for unmanaged ERP since we will lazily create
queues as data is enqueued rather than creating all queues during
storage initialization. This means that the FAST_BATCH priority queue
which checks for the key with a frequency of 1 second, may not exist for
some time. But we need the key in order to write data, so we move this
periodic key request/update logic to storage instead.

BUG=b:237809917
TEST=FEATURES="test" emerge-$BOARD missive

Change-Id: I4622580eb0064c4170e7d2bee190e2e7eaac9601
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4361659
Commit-Queue: Josh Hilke <jrhilke@google.com>
Reviewed-by: Hong Xu <xuhong@google.com>
Reviewed-by: Leonid Baraz <lbaraz@chromium.org>
Tested-by: Josh Hilke <jrhilke@google.com>