hiberman: Use TPM derived keys for hiberintegrity volume.
The TPM derived key is written to a tmpfs file, similar to how
biod handles tpm derived keys. This change will load and immediately
unlink the tpm derived key.

The other side of this is in mount_encrypted:
https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4261697

We use a ResumeSecretSeed from dbus to ensure the memory is zeroed when
it's dropped, we also rename this to HibernateKey as it can be used as
a container for the tpm key and the user pw derived key.

BUG=b:266012455
TEST=Tested on brya-hibernate

Change-Id: Icb65cdd44090fe6e09284fb53cb0b26951d4131e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4362479
Tested-by: Brian Geffon <bgeffon@chromium.org>
Reviewed-by: Allen Webb <allenwebb@google.com>