cryptohome: Wrap file paths in Platform logs with double quotes
cryptohome::Platform is used by arcvm-data-migrator, which redacts file
paths in the system logs to avoid disclosing PIIs to other users on the
same device. Since white spaces are more likely to be contained in the
file paths than double quotes or colons, the redaction logic is
implemented in such as way that white spaces following a file path is
parsed as the part of the file path, while double quotes or colons are
not (https://crrev.com/c/4338983/comment/ad9399b3_3d4df2b1/).
This CL double-quotes the file paths in the logs in cryptohome::Platform
unless they are not at the end of the logs. This CL only changes the
logs from methods used by arcvm-data-migrator.

BUG=b:251764421
TEST=FEATURES=test emerge-hatch cryptohome

Change-Id: Id16d3269583a8ceaa86c32b7ca362c484b84a8bb
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4365499
Reviewed-by: Daniil Lunev <dlunev@chromium.org>
Commit-Queue: Youkichi Hosoi <youkichihosoi@chromium.org>
Tested-by: Momoko Hattori <momohatt@chromium.org>
Reviewed-by: Youkichi Hosoi <youkichihosoi@chromium.org>