metrics: Create DestroyRecorderForTest method
SetRecorderForTest method allows to use a MockRecorder in tests
during SetUp. The singleton is only destroyed after the process
ends.

Expectations are evaluated after the mock object is destroyed.

DestroyRecorderForTest allows to ensure to free the recorder
during TearDown to ensure the expectations are run and to avoid
leaking the mock object.

BUG=None
TEST=1) FEATURES=test emerge-amd64-generic metrics
     2) Local test. Verified that the error message informing
        the expectations are not run and the mock object is leaked
	is gone using the method in TearDown.

Change-Id: I693180f0c15938f1fbd82572443e4744c04176c8
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4365079
Commit-Queue: Cristina Guerrero <crisguerrero@chromium.org>
Tested-by: Cristina Guerrero <crisguerrero@chromium.org>
Reviewed-by: Tony Yeoman <tby@chromium.org>
Reviewed-by: Helen Huang <helenhuang@chromium.org>