vm_tools: concierge: skip stateful disk update if a VM is suspended
If a VM is already suspended the stateful disk size update will not be
able to propagate and will finally timed out returning
"HandleStatefulUpdate RPC failed" error. Lets skip the stateful disk
update propagation for VMs which are suspended.

This patch actually fixes the graphic freeze after one of system
suspend/resume cycle on aurora-borealis when the Steam is running. The
problem is that the stateful disk update is currently triggered every 5s
and if hits at the time when the VM is already suspended (or not yet
resumed) it will block the concierge for 10s (due to
kDefaultTimeoutSeconds).

Since the concierge could block (which could even happen twice in a row)
for a while in a period after system resumes and before next system
suspend (suspend_stress_test), concierge, which should react on
SuspendDone and SuspendImminent, will not be able to do that. After
concierge finally unblocks it could have not enough time to actually
notify VM about suspend done or next suspend imminent (entire suspend
process has 20s timeout: kDefaultMaxSuspendDelayTimeout).

BUG=b:262518931
TEST=Run suspend_stress_test on aurora-borealis when Steam is running a
game. This test survives hundreds of cycles now. Without this patch the
graphic freeze and pass-through device were left in a bad state after
several cycles (usually less than 10).

Change-Id: Ic48177be661a0be9ccb32f5d4e870e730f21580a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4370478
Tested-by: Grzegorz Jaszczyk <jaszczyk@google.com>
Reviewed-by: Dmitry Torokhov <dtor@chromium.org>
Reviewed-by: Nic Hollingum <hollingum@google.com>
Reviewed-by: Daniel Ng <danielng@google.com>