hiberman: Get rid of ActiveMount
The purpose of ActiveMount is to keep track of a mounted filesystem
and umount it when the ActiveMount object is dropped. At this point
ActiveMount is only used in setup_hibermeta_lv(), where it is
configured to *not* unmount the file system on drop. The means the
ActiveMount is just (ab)used to mount the file system. That is
confusing, get rid of ActiveMount and add an API for mounting and
unmounting a filesystem instead.

BUG=b:267968644
TEST='hibermeta' volume still gets mounted

Change-Id: I265530a4d9121619e216cb68101dd6cbe2a20099
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4394249
Tested-by: Matthias Kaehlcke <mka@chromium.org>
Reviewed-by: Brian Geffon <bgeffon@chromium.org>
Commit-Queue: Matthias Kaehlcke <mka@chromium.org>