kerberos: Add source test plan to DIR_METADATA
As mentioned in https://crrev.com/c/4341233, this seems to be the
default set of test rules, which got added to many projects last year.
This can be adjusted later, if we notice that it includes too much
coverage, or not enough coverage.

See go/cros-cq-test-config-quickstart for more info.

BUG=None
TEST=Run CQ

Change-Id: I4896c94c54f6462280249f259b64b05b8d28a911
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4370880
Tested-by: Felipe Andrade <fsandrade@chromium.org>
Reviewed-by: Roman Sorokin <rsorokin@google.com>
Commit-Queue: Felipe Andrade <fsandrade@chromium.org>