diagnostics: Change cpu_stress to use V2 adapter in diagnostics service
Change the implementation of cpu_stress in diagnostics service to use
V2 adapter.

After this change, the CPU stress routine will attempt to run
stressapptest but leaving at least 500MiB of memory for the system.
stressapptest needs to run with at least 128MiB, so a total of 628MiB
is needed for the routine to run.

If the routine does not have over 628MiB memory, the routine will
have a status of kError.

BUG=b:182964589
TEST=(cr) FEATURES=test USE="-cros-debug" emerge-volteer diagnostics
     (dut) cros-health-tool diag --action=run_routine --routine=cpu_stress --cpu_stress_length_seconds=3

Change-Id: Ie73dff30600990fe9205446b469ae585e27c0ccd
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4297499
Tested-by: Ethan Cheng <yycheng@google.com>
Commit-Queue: Ethan Cheng <yycheng@google.com>
Reviewed-by: Wei-Luan Wang <weiluanwang@google.com>
Reviewed-by: Jimmy Gong <jimmyxgong@chromium.org>