chromeos-config: Map recovery input to keyboard for detachables
The mapping table is used to return the recovery input method based on
the form factor. The value POWER_BUTTON was designed for Wilco devices
(sarien, drallion), where their EC is not trusted. For detachables,
although the Power button is used in practice to confirm physical
presence, the field should be set to KEYBOARD (we should probably
rename it to DEFAULT to avoid confusion). Otherwise the instructions
in the firmware screen would be incorrect.

BUG=b:275033700
TEST=./regen.sh
TEST=See `Confirm` button when making transition to dev mode in FW
     screen

Cq-Depend: chromium:4374509
Change-Id: I09061f00f5a8cc723d25fbab2817c19c593f101d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4372720
Tested-by: Yidi Lin <yidilin@chromium.org>
Commit-Queue: Yidi Lin <yidilin@chromium.org>
Reviewed-by: Yu-Ping Wu <yupingso@chromium.org>
Reviewed-by: Jack Rosenthal <jrosenth@chromium.org>
Reviewed-by: Knox Chiou <knoxchiou@chromium.org>