rmad: Refine TryGetNextStateAtBoot
Allow RMAD_ERROR_OK return code when staying at the same state. This
could happen when we attempt to transition to the next state but gets
rejected. Sometimes this is an expected behavior and we should not
return an error code that propagates to metrics as a false alarm.

Also re-implement some |TryGetNextStateAtBoot| functions to avoid
re-using the existing |GetNextState| functions and return unnecessary
error codes.

BUG=b:271971970
TEST=FEATURES=test emerge-octopus rmad

Change-Id: Id0bbc856750c3e5e990117f11dcbd7da7c977065
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4381036
Tested-by: Cheng-Han Yang <chenghan@chromium.org>
Reviewed-by: Jeff Lin <jeffulin@google.com>
Commit-Queue: Cheng-Han Yang <chenghan@chromium.org>