patchpanel: add "domain-search" DHCP option for tethering
This CL adds the "domain-search" DHCP option to dnsmasq. With this CL,
the patchpanel could pass the list of the domain search to the DHCP
server at the downstream.

Note that currently the patchpanel always passes an empty list to
dnsmasq. In the following CL, the patchpanel should query the domain
search list from shill first, then pass it to dnsmasq.

BUG=b:275278561
TEST=FEATURES=test emerge-hatch patchpanel

Change-Id: I2e3fa2b96f96783a1ce168e694ee4324c86babff
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4374284
Reviewed-by: Jie Jiang <jiejiang@chromium.org>
Tested-by: Chih-Yu Huang <akahuang@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Commit-Queue: Chih-Yu Huang <akahuang@chromium.org>