cryptohome: Convert crc8 to C++
This isn't really that important or much of a change, but having a C
file makes this function a little weird. Converting it to C++ and
wrapping it in a namespace makes including and using it slightly more
normal. Also, doesn't hurt to add a few unit tests.

Also, combine the crc8 and crc32 files because there's no real need for
both of these to be separate modules. They're just a couple of short
functions, you're not saving much by breaking them up into separate
headers.

BUG=None
TEST=FEATURE=test emerge-$BOARD cryptohome

Change-Id: I126a02ec3f476e724b07403f7368ee78581c0db4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4382884
Tested-by: John Admanski <jadmanski@chromium.org>
Commit-Queue: John Admanski <jadmanski@chromium.org>
Reviewed-by: Zi Lin <lziest@google.com>