shill: wifi: update log level of debug level changes
This is generally useful information since it helps us know whether an
event was missing or whether it's not there because the log level is not
verbose enough. Log level changes don't occur frequently, so this should
be only a small number of additional log lines.

Also remove the minimal number of log checks in wifi_cqm_test to make
the unit tests pass. Note that the removed log lines are all redundant
checks.

BUG=b:270746800
TEST=`FEATURES=test emerge-brya shill`

Change-Id: I42136903e812afe57130ce0d3b942a825af06536
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4379101
Tested-by: Matthew Wang <matthewmwang@chromium.org>
Reviewed-by: Jintao Lin <jintaolin@chromium.org>
Reviewed-by: Kevin Lund <kglund@google.com>
Commit-Queue: Matthew Wang <matthewmwang@chromium.org>