diagnostics: Refactor unit test for memory routine V2
Refactor the unit test for memory routine V2 by changing variable
naming and reorganizing the code structure.

We split the MemoryRoutineV2Test into a base class and a derived class
to prepare for the adapter test as seen in CL:4356997.

We also add code to verify that memtester is ran with the expected
memory size.

BUG=b:276085206
TEST=(cr)FEATURES=test P2_TEST_FILTER='*' emerge-volteer diagnostics

Change-Id: I6a56decf316506817b0d7c7c9b15568e2731ef6c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4380317
Reviewed-by: Chung-sheng Wu <chungsheng@google.com>
Tested-by: Ethan Cheng <yycheng@google.com>
Reviewed-by: Wei-Luan Wang <weiluanwang@google.com>
Commit-Queue: Ethan Cheng <yycheng@google.com>