libsegmentation: Library to know what software feature to enable
Boiler plate code for the feature_management library.

Define an interface and main object that define the interface.
A fake implementation is provided for unit test.
The real implementation is empty for now.

"Inspired" from libcrossystem.

To use:
 Declare FeatureManagementImpl object in the main code.
 In the test code, declare a FeatureManagementFake.
 Have FeatureManagement feature_management_ object in the constructor
 of the main object, that can be instantiated either by the real or the
 fake feature_management object.

This is unrealled to featured. Although both returns if a given feature
is enabled or not, featured uses the finch framework while
feature_managent uses a local store.

BUG=b:244210433
TEST=unit tests.

Cq-Depend: chromium:4380482
Change-Id: I4b3752f58c6bf310bfe13b0763771e71cfc5993b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4380817
Tested-by: Gwendal Grignou <gwendal@chromium.org>
Commit-Queue: Gwendal Grignou <gwendal@chromium.org>
Reviewed-by: Sarthak Kukreti <sarthakkukreti@google.com>