cryptohome: data_migrator: Do not carry EOFBLOCKS forward
Older kernels use a flag EXT4_EOFBLOCKS_FL to help older e2fsprogs
cleaning the filesystem. This flag is deprecated (removed in 5.10) and
in 5.4 we are not allowed to set it. (The kernel will return EOPNOTSUPP).
It can linger around in the filesystem if the machine was updated from
an old kernel.
In consequence, when we see this flag set during migration, omit it when
setting flags to the new location.

BUG=b:274833188
TEST=`P2_TEST_FILTER="MigrationHelperTest.*" USE="${GG_USE}"
FEATURES=test emerge-brya cryptohome`

Change-Id: Ice9892c26d2fd8607b71393051be2356c118bd29
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4380858
Reviewed-by: Youkichi Hosoi <youkichihosoi@chromium.org>
Reviewed-by: Momoko Hattori <momohatt@chromium.org>
Commit-Queue: Gwendal Grignou <gwendal@chromium.org>
Tested-by: Gwendal Grignou <gwendal@chromium.org>