cryptohome: data_migrator: Simplify ExtFileAttributes test
The unit tests are using FakePlatform, which limits interaction with the
host filesystem. In particular SetExtFileAttributes() save the flags
internally, and GetExtFileAttributes() retrieve from internal data
structure. In consequence, we are guarantee that what we set is what we
get, without the host filesystem interfering.

Simplify the test in consequence.

BUG=b:266768936, b:274833188
TEST=`P2_TEST_FILTER="MigrationHelperTest.*" USE="${GG_USE}"
FEATURES=test emerge-brya cryptohome`

Change-Id: I2bf66e1fcf0d4f76d8e5f1bdfdef08017c941a9e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4380857
Commit-Queue: Gwendal Grignou <gwendal@chromium.org>
Reviewed-by: Youkichi Hosoi <youkichihosoi@chromium.org>
Tested-by: Gwendal Grignou <gwendal@chromium.org>