shill: Simplify DeviceClaimer logic (part 3)
This patch removes the logic ManagerDBusAdaptor to track the DBus
service who claims the devices. Since ClaimInterface is a test-only
interface now and we are not really using in the current tast test
codes (the names we're passing in now are not really D-Bus service
names), we should be able to remove this code.

Note that this patch will affect the behavior of the ClaimInterface
method:
- Previously, in tast test, the claimed interface will be released
  immediately after the call to ClaimInterface.
- Now, in tast test, the claimed interface will not be managed by shill
  until ReleaseInterface is called, or shill restarts.

The new behavior should be the expected one.

BUG=b:275481908
TEST=FEATURES="test" emerge-${BOARD} shill

Change-Id: I8bb46e8714e0e95438dea11d7b21f4ea05d44086
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4383631
Commit-Queue: Jie Jiang <jiejiang@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Tested-by: Jie Jiang <jiejiang@chromium.org>
Reviewed-by: Damien Dejean <damiendejean@google.com>