shill: fix supplicant properties for Passpoint with TLS
To match a network with no Organisation Identifier (OI), supplicant
needs to check the access point realm against the realm specified in the
credentials, but also the TLS credentials validity. Until now we were
not providing the username which lead to a match failure.

This CL adds the username to supplicant properties when the credentials
use EAP-TLS.

BUG=b:276188696
TEST=tast run dut wifi.PasspointSelection.* wifi.PasspointRoaming.*
TEST=FEATURES="test" emerge-${BOARD} chromeos-base/shill

Disallow-Recycled-Builds: test-failures
Change-Id: If5825a9d786860070942dc6130eccabc15c4e1b9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4382006
Reviewed-by: Jason Iman <jasongustaman@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Commit-Queue: Damien Dejean <damiendejean@google.com>
Tested-by: Hugo Benichi <hugobenichi@google.com>