shill: cellular: limit the number of GID digits in metrics
Because the GID1 value in structured metrics is of type int64, the
largest hex value that can be stored is 0x7FFFFFFFFFFFFFFF and any
value larger than that fails to parse. Change the number of Hex
digits from 16 to 15 to fit into an int.

BUG=b:276311050
TEST=Unit tests.

Change-Id: I9bc3e06c91a7b4c6f899e4d1e64665e5ee2ef767
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4386074
Tested-by: Andrew Lassalle <andrewlassalle@chromium.org>
Commit-Queue: Andrew Lassalle <andrewlassalle@chromium.org>
Auto-Submit: Andrew Lassalle <andrewlassalle@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>