cryptohome: Add support for delay-based PINs
This adds support to the pinweaver auth block for creating PIN factors
using the new gradual backup delay schedule. The feature is guarded
behind the PIN migration flag and so is still disabled by default.

This also inserts some of the plumbing for implementing migration on a
derive, but leaves the actual implementation as a TODO for now because
it's not quite trivial to implement compared to the Create.

BUG=b:272560921
TEST=FEATURES=test emerge-$BOARD cryptohome

Change-Id: Ia9c840cc3b6cdab2e5d2419316ed8b3261e7d8a5
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4385805
Reviewed-by: Hardik Goyal <hardikgoyal@chromium.org>
Tested-by: Hardik Goyal <hardikgoyal@chromium.org>
Tested-by: John Admanski <jadmanski@chromium.org>
Commit-Queue: John Admanski <jadmanski@chromium.org>