secanomalyd: migrate to procfs process enumeration
Implements a new way for enumerating system processes, which uses
procfs. This is needed in order to read process attributes that are not
exposed through ps, such as seccomp status of the process, which is
found inside the process status file.
This is achieved by traversing process sub-directories under /proc and
reading the required attributes from various files and symlinks.
Adds several newly introduced attributes to the list of existing
properties for the ProcEntry object. Adds test coverage for the new
functionality and updates the reporter test accordingly.
Removes the ps-based process enumeration logic along with the tests.

BUG=b:278131121
TEST=Correctly lists all the processes on the system and the correct
attributes for each.
TEST=Crash data includes the list of args for all the processes similar
to before.
TEST=Correctly applies the process filters.
TEST=All tests pass.

Change-Id: I1b4ac312b6b7b3be9499ca0be9e2b2835ea965e6
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4387279
Commit-Queue: Shervin Oloumi <enlightened@chromium.org>
Tested-by: Shervin Oloumi <enlightened@chromium.org>
Reviewed-by: Jorge Lucangeli Obes <jorgelo@chromium.org>