hiberman: lvm: Rename get_active_lvs() to get_lvs()
The function get_active_lvs() in the lvm module is misnamed,
it returns a list of all LVs, not only the active ones. Rename
it to get_lvs() to avoid confusion.

Also add a parameter to specify the volume group, which is
consistent with the rest of the lvm APIs.

BUG=b:267968644
TEST=hibernate still works on brya

Change-Id: Idaf167680289f61e8707b50badde54eaf5a61bfb
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4394250
Tested-by: Matthias Kaehlcke <mka@chromium.org>
Reviewed-by: Brian Geffon <bgeffon@chromium.org>
Commit-Queue: Matthias Kaehlcke <mka@chromium.org>