trunks: libhwsec: Make start_bio_auth take Blob fields
The start_bio_auth takes a nonce, and returns a nonce, an IV, and an
encrypted secret. All of these aren't sensitive data, so should use Blob
instead of SecureBlob. The libhwsec interface will be modified in the
follow-up CL, currently it is changed to do the transformation when
calling that trunks API.

BUG=b:247704971
TEST=libhwsec/trunks unittests

Change-Id: If9bf1c0e7e572b68b6324ac27135189775021bfd
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4387223
Tested-by: Howard Yang <hcyang@google.com>
Commit-Queue: Howard Yang <hcyang@google.com>
Reviewed-by: Yi Chou <yich@google.com>