libhwsec: cryptohome: Make StarBioAuth take Blob fields
The StartBioAuth takes a nonce, and returns a nonce, an IV, and an
encrypted secret. All of these aren't sensitive data, so should use Blob
instead of SecureBlob. The LE cred manager interface will be modified in the
follow-up CL, currently it is changed to do the transformation when
calling that libhwsec API.

BUG=b:247704971
TEST=cryptohome/libhwsec unittests

Change-Id: Idb6f71b6563b81f858844eca92cd86e3ecdb8e38
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4387225
Tested-by: Howard Yang <hcyang@google.com>
Reviewed-by: Yi Chou <yich@google.com>
Commit-Queue: Howard Yang <hcyang@google.com>