lorgnette: Remove unnecessary locking
When lorgnette was first written, we added locking around scanner
accesses because we thought it was possible for d-bus calls to be called
concurrently.  It turns out that DBusDaemon unconditionally uses a
SingleThreadTaskExecutor, so this can never happen.  Drop the locks in
favor of a sequence checker so we can simplify the code.

BUG=b:275283459
TEST=tast run ("paper-io_scanning")

Change-Id: If0084ed972849fb6a2fb5ff5a3c233046716cc78
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4390944
Commit-Queue: Paul Moy <pmoy@chromium.org>
Tested-by: Benjamin Gordon <bmgordon@chromium.org>
Auto-Submit: Benjamin Gordon <bmgordon@chromium.org>
Reviewed-by: Paul Moy <pmoy@chromium.org>