shill: vpn: Set gateway address to any for wireguard
In shill, there is no difference between empty and "0.0.0.0"/"::" for
the gateway address of a VPN connection, but the current shill users
will be unhappy to see an empty value for the gateway property.

For a WireGuard service, the system UI used to set the gateway value to
empty, but that logic was removed in https://crrev.com/c/4225076 for the
IPv6 overlay support. This causes some regression for the WireGuard
service.

BUG=b:276506661
TEST=FEATURES="test" emerge-${BOARD} shill
TEST=tast run $DUT network.VPNRouting.wireguard*
  network.VPNConnect.wireguard*
TEST=manually checked that with this patch, dnsproxy is able to get the
  nameservers according to the log

Change-Id: Ib73d5592d67d4ad75dcd1998777c21dc66ae85c1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4391683
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Commit-Queue: Jie Jiang <jiejiang@chromium.org>
Tested-by: Jie Jiang <jiejiang@chromium.org>