printscanmgr: change D-Bus interface to take protos
Change all of printscanmgr's D-Bus methods to both accept and return a
protobuf. This will make interfaces changes easier in the future.

This includes one significant change to PrintscanDebugSetCategories -
now that we're using a proto, we can add a new field to explicitly ask
printscanmgr to disable log collection, instead of relying on an empty
categories vector to have a special meaning (disabling logging).

BUG=b:257070388
TEST=FEATURES=test emerge-nami printscanmgr

Change-Id: I67359cfda98cc11328c0f81ee91124a820be717f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4395888
Reviewed-by: Benjamin Gordon <bmgordon@chromium.org>
Commit-Queue: Paul Moy <pmoy@chromium.org>
Tested-by: Paul Moy <pmoy@chromium.org>
Reviewed-by: Mike Frysinger <vapier@chromium.org>