installer: Run part of reven partition migration during updates
Refactor the migration so that the planning stage is run for both
installs and updates. This stage only reads information, it doesn't
modify anything, so a failure there should not prevent postinstall from
succeeding.

If an error does occur during planning, or if the migration is not
needed, a metric is sent.

The actual migration is still only run during install.

BUG=b:249809975
TEST=cros_workon_make --test chromeos-installer

Change-Id: I0a234f7b9e1a3c7fa1d65994b9739d7230af705c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4395283
Reviewed-by: Ted Brandston <tbrandston@google.com>
Commit-Queue: Nicholas Bishop <nicholasbishop@google.com>
Tested-by: Nicholas Bishop <nicholasbishop@google.com>