crosh: wireguard command no longer sets IPAddress in StaticIPConfig
Setting and reading the IP address from StaticIPConfig is no longer
needed, because 1) shill already does a migration for the IP address
property and 2) Chrome no longer reads and writes this property.

This also fixes the issue that dns-proxy may not work for the wg service
created by the wireguard command in shill because we do not set the
Prefixlen together.

BUG=b:276849438
TEST=FEATURES="test" emerge-${BOARD} crosh
TEST=`tast run $DUT network.WireguardCrosh` and checked that there is no
  log about "empty Prefixlen" in dns-proxy.

Change-Id: I25078c70594279a909c7533a5246b9057d98e4a9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4397126
Commit-Queue: Jie Jiang <jiejiang@chromium.org>
Tested-by: Jie Jiang <jiejiang@chromium.org>
Reviewed-by: Taoyu Li <taoyl@chromium.org>
Reviewed-by: Allen Webb <allenwebb@google.com>