hiberman: Make thinpool writable before tearing down LVs
During the resume-init phase of a bootstrap boot hiberman activates
the thinpool in read-only mode if a resume fro hibernate is pending.
This is done to avoid metadata changes that could result in
inconsistencies in the to-be-resumed hibernated world, which
wouldn't be aware of these changes.

If the system can't resume from a hibernate image we tear down
existing DMs and LVs related with the 'hiberimage', to re-create
them with the keys of the user that is currently logged in. In
order to do that the thinpool needs to be writable, so change it
back to read-write mode.

BUG=b:266012455
TEST=no failures to remove LVs when resume is aborted

Change-Id: I357740d27f6fba217d8f789e26a75e69b54e06bf
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4394251
Reviewed-by: Brian Geffon <bgeffon@chromium.org>
Commit-Queue: Matthias Kaehlcke <mka@chromium.org>
Tested-by: Matthias Kaehlcke <mka@chromium.org>