hiberman: Move DM snapshot COW files to tmpfs
Currently the DM snapshot COW files are located on the 'hibermeta'
volume. As a result the volume can not be unmounted before switching
to the hibernated image. This complicates the transition of the
'hibermeta' file system from ext4 to ext2, since we want to cleanly
unmount the non-journaling file system whenever possible to avoid
corruption.

The DM snapshots are exclusively used by the bootstrap boot, so it's
not really necessary to host the COW files on persistent storage. Use
tmpfs for these temporary files.

BUG=b:267968644
TEST=hibernate still works on brya

Change-Id: I1a9fb21e42bf52e708d63c6078e2ecc400e3633b
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4401960
Tested-by: Matthias Kaehlcke <mka@chromium.org>
Reviewed-by: Brian Geffon <bgeffon@chromium.org>
Commit-Queue: Matthias Kaehlcke <mka@chromium.org>